Puravankara Cuffe Parade Mumbai
Quality residential space is now easily obtainable if you live in Mumbai, thanks to Puravankara Group the renowned residential builders. Contemporary lifestyle combined with the need to live in clean and airy surroundings has become a necessity for people looking for new accommodations in Mumbai ca go for Puravankara Cuffe Parade being developed in Mumbai, fulfils that need to the ultimate limits.

It is a modern architecture designed by the most famous architects who are in trend with the modern requirements of the contemporary lifestyle. Apartments ranging from 2BHK, 3BHK to 4BHK of different configurations are being offered by the prestigious development and the prices. The project is conceived to be built in an area of 8.7 sprawling acres and 80% of the total land is dedicated to open space. You can imagine how this is going to be beneficial for the residents who decide to take up tenancy. Puravankara Cuffe Parade Mumbai with a maximum total area is going to be dedicated to a green environment consisting of parks, tasteful landscapes, a gazebo, gardens and intertwining pathways and joggers’ tracks.

The upcoming Puravankara Cuffe Parade is a masterpiece in terms of design and layout and numerous world-class amenities within its premises to residents. Some of the features provided by the developer as basic amenities are high-speed lifts, a 24 x 7 water supply, and power backup. A quick responding firefighting system is an added security for the resident’s safety and you will have ample parking space for your car and will have free access to the clubhouse and gym at all permitted times. a huge swimming pool is being built for both adults and kids and socializing at the clubhouse, community hall, and children's play area. Adequate open space has been dedicated to the solitary and gated community with guards.

You have seamless connectivity because the development is located on the brim of Nariman Point and you can access all necessary outside facilities and adjacent localities easily via wide roads. Don’t let the opportunity to live near the sea views because space could beat premium here! 
To know more - https://bit.ly/3uzUSat